Working(ish) dot grid
